TAJO-946: UniformRangePartition leads to out of range partitions

+++ core/src/main/java/tajo/engine/planner/UniformRangePartition.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2012 Database Lab., Korea Univ.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package tajo.engine.planner;
 
 import com.google.common.base.Preconditions;
@@ -14,9 +30,6 @@ import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.List;
 
-/**
- * @author Hyunsik Choi
- */
 public class UniformRangePartition extends RangePartitionAlgorithm {
   private int variableId;
   private BigDecimal[] cardForEachDigit;
@@ -279,7 +292,7 @@ public class UniformRangePartition extends RangePartitionAlgorithm {
           if (overflowFlag[i]) {
             end.put(i, DatumFactory.createLong(range.getStart().get(i).asInt() + incs[i].longValue()));
           } else {
-            end.put(i, DatumFactory.createLong(last.get(i).asInt() + incs[i].longValue()));
+            end.put(i, DatumFactory.createLong(last.get(i).asLong() + incs[i].longValue()));
           }
           break;
         case FLOAT:
